requirements.txt
blackmarblepy
geopandas
pandas
numpy
pyarrow
matplotlib
python-dotenv
requests